Fix bug with applying golang formatting
goimports should be run post writing to disk rather than pre writing,
otherwise goimports is not working on the changes in the buffer. This is
different to other languages which use the LSP - in that case the
formatting should be applied before writing to disk. This is because
goimports is run outside of neovim.