Redraw diagnostics after save
We automatically run formatting on save. This removes diagnostics.
Unless the LSP returns updated diagnostics then they are not redrawn.
This change forces all diagnostics to be redrawn straight after
formatting. There may be a danger that the diagnostics are not up to
date with the changes made by formatting. In that case I think the LSP
will send updated diagnostics. Presumably these are processed after the
old diagnostics have been written so they will overwrite them. Even if
they were somehow processed after we apply the redraw, we should be
redrawing the updated diagnostics.

This change also removes the timeout on applying formatting. It seems
like we would always want formatting to be applied. If for some reason
formatting is painfully slow then it suggests some other problem which
this change will highlight rather than hiding the problem like it did
previously.