Update CursorHold updatetime
The keyboard repeat delay is the time between when you press and hold a
key and when it starts repeating. If the CursorHold updatetime is less
than this then the CursorHold will trigger before the key starts
repeating. This will cause the git line info to run, causing a stutter.

Typically the keyboard delay is around 500ms so we set the CursorHold
updatetime to 600ms.